+++ src/muos/steps/arch_chroot/update_vconsole_conf.py
@@ -8,8 +8,6 @@ class UpdateVconsoleConf(Step):
     def main(self, environment: Environment) -> None:
         super().main(environment)
         vconsole_conf = Path(environment.mnt) / 'etc' / 'vconsole.conf'
-        lang = environment.main_locale.split(' ')[0]
         vconsole_conf.write_text('\n'.join([
             'KEYMAP={}'.format(environment.keymap),
-            'LANG={}'.format(lang),
         ]))